v0.8 Sprint 3 (QA-015): decompose CommunityHttpClientEngine — close GodClass + TooManyMethods
First Sprint 3 GodClass closure (Quality batch II follow-up after Sprint 1
closure of QA-010..014). Splits CommunityHttpClientEngine along two
cohesive HTTP/1.x seams:

  - Request byte layout (request line + headers + body) → encoder
  - Response parsing (status line + headers + body length) → decoder

Three files now:

  CommunityHttpClientRequestEncoder (153 LOC, pkg-private static)
    * writeRequest(seg, request, connection, bodyBytes) — top entry
    * writeRequestLine / writeRequestHeaders / writeDefaultRequestHeaders
    * writeRequestBody / writeAscii / versionToken
    * HeaderPresence + HeaderWriteResult records
    * Default headers: Host: host:port (from TransportConnection),
      Content-Length: N, Connection: close (when not user-provided)

  CommunityHttpClientResponseDecoder (220 LOC, pkg-private static)
    * decodeResponse(allocator, aggregate, total, requestVersion) — top entry
    * resolveHeaderTerminator / resolveExpectedTotal / isResponseComplete
      — incremental completeness helpers consulted by the engine's read loop
    * parseHeaders / parseStatusLine / resolveBodyLength /
      resolveContentLengthFromHeaders / findHeaderTerminator / asciiString
    * StatusLine record
    * Class-level @SuppressWarnings("PMD.CyclomaticComplexity") with
      rationale (status line + header parsing has flat CC)

  CommunityHttpClientEngine (219 LOC, refactored — public surface unchanged)
    * Public surface UNCHANGED: HttpClientEngine.start/send/close/isRunning/
      engineName all preserved verbatim
    * Per-call orchestration (transport.connect + stream.openStream +
      sendRequest + readResponse) retained
    * sendRequest now allocates outbound buffer + delegates byte layout
      to CommunityHttpClientRequestEncoder.writeRequest
    * readResponse retains the multi-state read loop (incremental header
      terminator + expected total + completeness) but delegates each
      step to CommunityHttpClientResponseDecoder
    * Dependency resolution (allocator + transport) retained at the bottom
    * ResolvedHttpClientDeps internal record retained
    * Single class-level @SuppressWarnings("PMD.CyclomaticComplexity")
      remains — the read loop is intrinsically cohesive (5 state vars
      mutated together in a single while body; extracting would require
      a holder record with no functional benefit)

Suppression delta on CommunityHttpClientEngine:

  Removed (closed, 2):
    - PMD.GodClass
    - PMD.TooManyMethods

  Retained (1):
    - PMD.CyclomaticComplexity (rationale inline)

  New on extracted helpers (with inline rationale):
    - CommunityHttpClientResponseDecoder: PMD.CyclomaticComplexity
      (flat status+headers+completeness parse CC)
    - CommunityHttpClientRequestEncoder: no suppressions needed

Delta:
  CommunityHttpClientEngine 470 → 219 LOC (-251 / -53%)
  + CommunityHttpClientRequestEncoder 153 LOC
  + CommunityHttpClientResponseDecoder 220 LOC
  Class-level suppressions on engine: 3 → 1 (-2)

Test plan:
  - CommunityHttpClientEngineTckTest 7/7 green (TCK contract preserved)
  - AbstractHttpClientEngineTck nested classes 7/7 green
    (Identity 2/2, StartState 2/2, CreatedState 2/2, Close 2/2)
  - ExerisWebClientIntegrationTest 8/8 green (the Sprint 2 client SPI
    that layers on top of CommunityHttpClientEngine — no regression)
  - Full reactor verify SUCCESS with -P coverage (Sprint 6 JaCoCo gate
    still passes: community 69% > 0.55 floor)
  - PMD 0 violations on community module
  - Checkstyle 0 violations on community module

Sprint 3 GodClass tracker after this PR:

| QA-* | Class                          | LOC          | Class-suppressions |
| QA-015 | CommunityHttpClientEngine    | 470→219      | 3→1 (-2)           |
| QA-016 | NativeTcpStream (HEUR-061)   | 1229         | next               |
| QA-017 | JdbcFlowSnapshotStore        | TBD          | next               |
| QA-018 | CommunityHttp2SessionProcessor + SubsystemOrchestrator | next |
